+++ resolwe/permissions/tests/test_data.py
@@ -216,7 +216,7 @@ class DataTestCase(ResolweAPITestCase):
                                     'descriptor_schema', 'descriptor', 'id', 'process_slug', 'process_name',
                                     'process_input_schema', 'process_output_schema', 'size',
                                     'current_user_permissions', 'descriptor_dirty', 'tags', 'process_memory',
-                                    'process_cores'])
+                                    'process_cores', 'collections', 'entities'])
 
         # user w/ public permissions
         resp = self._get_detail(1, self.user2)
@@ -227,7 +227,7 @@ class DataTestCase(ResolweAPITestCase):
                                     'descriptor_schema', 'descriptor', 'id', 'process_slug', 'process_name',
                                     'process_input_schema', 'process_output_schema', 'size',
                                     'current_user_permissions', 'descriptor_dirty', 'tags', 'process_memory',
-                                    'process_cores'])
+                                    'process_cores', 'collections', 'entities'])
 
     def test_get_detail_no_perms(self):
         # public user w/o permissions