+++ docs/CHANGELOG.rst
@@ -20,6 +20,8 @@ Changed
 Added
 -----
 - Add support for URLs in ``basic:file:`` fields in Django tests
+- Add ``collections`` and ``entities`` fields to Data serializer, with optional
+  hydration using ``hydrate_collections`` and/or ``hydrate_entities``
 
 Fixed
 -----